requirements.txt
# MNQ Bot – same as mnq_bot/requirements.txt (for Railway build from repo root)
python-telegram-bot>=20.0,<21.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
matplotlib>=3.7.0
APScheduler>=3.10.0
requests>=2.28.0
beautifulsoup4>=4.12.0
flask>=3.0.0
gunicorn>=21.0.0
websocket-client>=1.6.0
pytz>=2023.3
python-dotenv>=1.0.0
